requirements.txt
django
pillow
weasyprint
gunicorn
djangorestframework
uvicorn[standard]>=0.27.1